Move shared methods to JoinSwappingUtils
Move methods shared between multiple join swapping optimizers to
JoinSwappingUtils.
