+++ presto-main/src/main/java/com/facebook/presto/sql/planner/iterative/rule/DetermineJoinDistributionType.java
@@ -25,13 +25,11 @@ import com.facebook.presto.spi.plan.PlanNode;
 import com.facebook.presto.spi.plan.TableScanNode;
 import com.facebook.presto.spi.plan.ValuesNode;
 import com.facebook.presto.sql.analyzer.FeaturesConfig.JoinDistributionType;
-import com.facebook.presto.sql.planner.iterative.GroupReference;
 import com.facebook.presto.sql.planner.iterative.Lookup;
 import com.facebook.presto.sql.planner.iterative.Rule;
 import com.facebook.presto.sql.planner.optimizations.PlanNodeSearcher;
 import com.facebook.presto.sql.planner.plan.JoinNode;
 import com.facebook.presto.sql.planner.plan.RemoteSourceNode;
-import com.facebook.presto.sql.planner.plan.UnnestNode;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Ordering;
@@ -39,27 +37,25 @@ import io.airlift.units.DataSize;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Stream;
 
 import static com.facebook.presto.SystemSessionProperties.getJoinDistributionType;
 import static com.facebook.presto.SystemSessionProperties.getJoinMaxBroadcastTableSize;
 import static com.facebook.presto.SystemSessionProperties.isSizeBasedJoinDistributionTypeEnabled;
 import static com.facebook.presto.cost.CostCalculatorWithEstimatedExchanges.calculateJoinCostWithoutOutput;
 import static com.facebook.presto.sql.analyzer.FeaturesConfig.JoinDistributionType.AUTOMATIC;
+import static com.facebook.presto.sql.planner.iterative.rule.JoinSwappingUtils.isBelowBroadcastLimit;
+import static com.facebook.presto.sql.planner.iterative.rule.JoinSwappingUtils.isSmallerThanThreshold;
 import static com.facebook.presto.sql.planner.optimizations.QueryCardinalityUtil.isAtMostScalar;
 import static com.facebook.presto.sql.planner.plan.JoinNode.DistributionType.PARTITIONED;
 import static com.facebook.presto.sql.planner.plan.JoinNode.DistributionType.REPLICATED;
 import static com.facebook.presto.sql.planner.plan.Patterns.join;
 import static java.lang.Double.NaN;
-import static java.lang.Double.isNaN;
 import static java.util.Objects.requireNonNull;
 
 public class DetermineJoinDistributionType
         implements Rule<JoinNode>
 {
     private static final Pattern<JoinNode> PATTERN = join().matching(joinNode -> !joinNode.getDistributionType().isPresent());
-    private static final List<Class<? extends PlanNode>> EXPANDING_NODE_CLASSES = ImmutableList.of(JoinNode.class, UnnestNode.class);
-    private static final double SIZE_DIFFERENCE_THRESHOLD = 8;
 
     private final CostComparator costComparator;
     private final TaskCountEstimator taskCountEstimator;
@@ -161,19 +157,6 @@ public class DetermineJoinDistributionType
         return getSyntacticOrderJoin(joinNode, context, AUTOMATIC);
     }
 
-    public static boolean isBelowBroadcastLimit(PlanNode planNode, Context context)
-    {
-        DataSize joinMaxBroadcastTableSize = getJoinMaxBroadcastTableSize(context.getSession());
-        return getSourceTablesSizeInBytes(planNode, context) <= joinMaxBroadcastTableSize.toBytes();
-    }
-
-    public static boolean isSmallerThanThreshold(PlanNode planNodeA, PlanNode planNodeB, Context context)
-    {
-        double aOutputSize = getFirstKnownOutputSizeInBytes(planNodeA, context);
-        double bOutputSize = getFirstKnownOutputSizeInBytes(planNodeB, context);
-        return aOutputSize * SIZE_DIFFERENCE_THRESHOLD < bOutputSize;
-    }
-
     public static double getSourceTablesSizeInBytes(PlanNode node, Context context)
     {
         return getSourceTablesSizeInBytes(node, context.getLookup(), context.getStatsProvider());
@@ -183,7 +166,7 @@ public class DetermineJoinDistributionType
     static double getSourceTablesSizeInBytes(PlanNode node, Lookup lookup, StatsProvider statsProvider)
     {
         boolean hasExpandingNodes = PlanNodeSearcher.searchFrom(node, lookup)
-                .whereIsInstanceOfAny(EXPANDING_NODE_CLASSES)
+                .whereIsInstanceOfAny(JoinSwappingUtils.EXPANDING_NODE_CLASSES)
                 .matches();
         if (hasExpandingNodes) {
             return NaN;
@@ -195,56 +178,6 @@ public class DetermineJoinDistributionType
 
         return sourceNodes.stream()
                 .mapToDouble(sourceNode -> statsProvider.getStats(sourceNode).getOutputSizeInBytes(sourceNode))
-                .sum();
-    }
-
-    private static double getFirstKnownOutputSizeInBytes(PlanNode node, Context context)
-    {
-        return getFirstKnownOutputSizeInBytes(node, context.getLookup(), context.getStatsProvider());
-    }
-
-    /**
-     * Recursively looks for the first source node with a known estimate and uses that to return an approximate output size.
-     * Returns NaN if an un-estimated expanding node (Join or Unnest) is encountered.
-     * The amount of reduction in size from un-estimated non-expanding nodes (e.g. an un-estimated filter or aggregation)
-     * is not accounted here. We make use of the first available estimate and make decision about flipping join sides only if
-     * we find a large difference in output size of both sides.
-     */
-    @VisibleForTesting
-    public static double getFirstKnownOutputSizeInBytes(PlanNode node, Lookup lookup, StatsProvider statsProvider)
-    {
-        return Stream.of(node)
-                .flatMap(planNode -> {
-                    if (planNode instanceof GroupReference) {
-                        return lookup.resolveGroup(node);
-                    }
-                    return Stream.of(planNode);
-                })
-                .mapToDouble(resolvedNode -> {
-                    double outputSizeInBytes = statsProvider.getStats(resolvedNode).getOutputSizeInBytes(resolvedNode);
-                    if (!isNaN(outputSizeInBytes)) {
-                        return outputSizeInBytes;
-                    }
-
-                    if (EXPANDING_NODE_CLASSES.stream().anyMatch(clazz -> clazz.isInstance(resolvedNode))) {
-                        return NaN;
-                    }
-
-                    List<PlanNode> sourceNodes = resolvedNode.getSources();
-                    if (sourceNodes.isEmpty()) {
-                        return NaN;
-                    }
-
-                    double sourcesOutputSizeInBytes = 0;
-                    for (PlanNode sourceNode : sourceNodes) {
-                        double firstKnownOutputSizeInBytes = getFirstKnownOutputSizeInBytes(sourceNode, lookup, statsProvider);
-                        if (isNaN(firstKnownOutputSizeInBytes)) {
-                            return NaN;
-                        }
-                        sourcesOutputSizeInBytes += firstKnownOutputSizeInBytes;
-                    }
-                    return sourcesOutputSizeInBytes;
-                })
                 .sum();
     }
